feat: 完善评分功能和热度查询

+++ src/main/java/plus/maa/backend/repository/entity/Copilot.java
@@ -50,7 +50,7 @@ public class Copilot implements Serializable {
     private int views;
 
     //热度
-    private int hotScore;
+    private double hotScore;
 
 
     //难度
